+++ hudi-common/src/test/java/org/apache/hudi/common/util/collection/TestBitCaskDiskMap.java
@@ -18,22 +18,25 @@
 
 package org.apache.hudi.common.util.collection;
 
-import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericRecord;
-import org.apache.avro.generic.IndexedRecord;
-
 import org.apache.hudi.avro.HoodieAvroUtils;
 import org.apache.hudi.common.model.HoodieAvroPayload;
 import org.apache.hudi.common.model.HoodieKey;
 import org.apache.hudi.common.model.HoodieRecord;
 import org.apache.hudi.common.model.HoodieRecordPayload;
 import org.apache.hudi.common.table.timeline.HoodieActiveTimeline;
+import org.apache.hudi.common.testutils.AvroBinaryTestPayload;
 import org.apache.hudi.common.testutils.HoodieCommonTestHarness;
 import org.apache.hudi.common.testutils.SchemaTestUtil;
 import org.apache.hudi.common.testutils.SpillableMapTestUtils;
+import org.apache.hudi.common.util.HoodieRecordSizeEstimator;
 import org.apache.hudi.common.util.Option;
+import org.apache.hudi.common.util.SpillableMapUtils;
 
+import org.apache.avro.Schema;
+import org.apache.avro.generic.GenericRecord;
+import org.apache.avro.generic.IndexedRecord;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -45,75 +48,55 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Random;
 import java.util.Set;
+import java.util.UUID;
+import java.util.stream.Collectors;
 
 import static org.apache.hudi.common.testutils.SchemaTestUtil.getSimpleSchema;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * Test the rocksDb based Map {@link SpillableRocksDBBasedMap}
- * that is used by {@link ExternalSpillableMap}.
+ * Tests dis based map {@link BitCaskDiskMap}.
  */
-public class TestSpillableRocksDBBasedMap extends HoodieCommonTestHarness {
+public class TestBitCaskDiskMap extends HoodieCommonTestHarness {
 
   @BeforeEach
-  public void setUp() {
+  public void setup() {
     initPath();
   }
 
   @Test
-  public void testSimpleInsertSequential() throws IOException, URISyntaxException {
-    SpillableRocksDBBasedMap rocksDBBasedMap = new SpillableRocksDBBasedMap(basePath);
+  public void testSimpleInsert() throws IOException, URISyntaxException {
+    BitCaskDiskMap records = new BitCaskDiskMap<>(basePath);
     List<IndexedRecord> iRecords = SchemaTestUtil.generateHoodieTestRecords(0, 100);
     ((GenericRecord) iRecords.get(0)).get(HoodieRecord.COMMIT_TIME_METADATA_FIELD).toString();
-    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, rocksDBBasedMap);
+    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, records);
 
-    // Ensure the number of records is correct
-    assertEquals(rocksDBBasedMap.size(), recordKeys.size());
     // make sure records have spilled to disk
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > 0);
-    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = rocksDBBasedMap.iterator();
+    assertTrue(records.sizeOfFileOnDiskInBytes() > 0);
+    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = records.iterator();
     List<HoodieRecord> oRecords = new ArrayList<>();
     while (itr.hasNext()) {
       HoodieRecord<? extends HoodieRecordPayload> rec = itr.next();
       oRecords.add(rec);
       assert recordKeys.contains(rec.getRecordKey());
     }
-    assertEquals(recordKeys.size(), oRecords.size());
-  }
-
-  @Test
-  public void testSimpleInsertRandomAccess() throws IOException, URISyntaxException {
-    SpillableRocksDBBasedMap rocksDBBasedMap = new SpillableRocksDBBasedMap(basePath);
-    List<IndexedRecord> iRecords = SchemaTestUtil.generateHoodieTestRecords(0, 100);
-    ((GenericRecord) iRecords.get(0)).get(HoodieRecord.COMMIT_TIME_METADATA_FIELD).toString();
-    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, rocksDBBasedMap);
-
-    // make sure records have spilled to disk
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > 0);
-
-    Random random = new Random();
-    for (int i = 0; i < recordKeys.size(); i++) {
-      String key = recordKeys.get(random.nextInt(recordKeys.size()));
-      assert rocksDBBasedMap.get(key) != null;
-    }
   }
 
   @Test
   public void testSimpleInsertWithoutHoodieMetadata() throws IOException, URISyntaxException {
-    SpillableRocksDBBasedMap rocksDBBasedMap = new SpillableRocksDBBasedMap<>(basePath);
+    BitCaskDiskMap records = new BitCaskDiskMap<>(basePath);
     List<HoodieRecord> hoodieRecords = SchemaTestUtil.generateHoodieTestRecordsWithoutHoodieMetadata(0, 1000);
     Set<String> recordKeys = new HashSet<>();
     // insert generated records into the map
     hoodieRecords.forEach(r -> {
-      rocksDBBasedMap.put(r.getRecordKey(), r);
+      records.put(r.getRecordKey(), r);
       recordKeys.add(r.getRecordKey());
     });
     // make sure records have spilled to disk
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > 0);
-    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = rocksDBBasedMap.iterator();
+    assertTrue(records.sizeOfFileOnDiskInBytes() > 0);
+    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = records.iterator();
     List<HoodieRecord> oRecords = new ArrayList<>();
     while (itr.hasNext()) {
       HoodieRecord<? extends HoodieRecordPayload> rec = itr.next();
@@ -126,13 +109,13 @@ public class TestSpillableRocksDBBasedMap extends HoodieCommonTestHarness {
   public void testSimpleUpsert() throws IOException, URISyntaxException {
     Schema schema = HoodieAvroUtils.addMetadataFields(getSimpleSchema());
 
-    SpillableRocksDBBasedMap rocksDBBasedMap = new SpillableRocksDBBasedMap<>(basePath);
+    BitCaskDiskMap records = new BitCaskDiskMap<>(basePath);
     List<IndexedRecord> iRecords = SchemaTestUtil.generateHoodieTestRecords(0, 100);
 
     // perform some inserts
-    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, rocksDBBasedMap);
+    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, records);
 
-    long fileSize = rocksDBBasedMap.sizeOfFileOnDiskInBytes();
+    long fileSize = records.sizeOfFileOnDiskInBytes();
     // make sure records have spilled to disk
     assertTrue(fileSize > 0);
 
@@ -143,16 +126,13 @@ public class TestSpillableRocksDBBasedMap extends HoodieCommonTestHarness {
         ((GenericRecord) updatedRecords.get(0)).get(HoodieRecord.COMMIT_TIME_METADATA_FIELD).toString();
 
     // perform upserts
-    recordKeys = SpillableMapTestUtils.upsertRecords(updatedRecords, rocksDBBasedMap);
+    recordKeys = SpillableMapTestUtils.upsertRecords(updatedRecords, records);
 
-    // On disk (rocksDb), upserts should actual replace the original records, and the size
-    // should not increase. But due to the way the bytes are accounted,
-    // sizeOfFileOnDiskInBytes only maintains the total
-    // bytes written, and will increase with upserts.
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > fileSize);
+    // upserts should be appended to the existing file, hence increasing the sizeOfFile on disk
+    assertTrue(records.sizeOfFileOnDiskInBytes() > fileSize);
 
     // Upserted records (on disk) should have the latest commit time
-    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = rocksDBBasedMap.iterator();
+    Iterator<HoodieRecord<? extends HoodieRecordPayload>> itr = records.iterator();
     while (itr.hasNext()) {
       HoodieRecord<? extends HoodieRecordPayload> rec = itr.next();
       assert recordKeys.contains(rec.getRecordKey());
@@ -168,8 +148,48 @@ public class TestSpillableRocksDBBasedMap extends HoodieCommonTestHarness {
   }
 
   @Test
+  public void testSizeEstimator() throws IOException, URISyntaxException {
+    Schema schema = SchemaTestUtil.getSimpleSchema();
+
+    // Test sizeEstimator without hoodie metadata fields
+    List<HoodieRecord> hoodieRecords = SchemaTestUtil.generateHoodieTestRecords(0, 1, schema);
+
+    long payloadSize =
+        SpillableMapUtils.computePayloadSize(hoodieRecords.remove(0), new HoodieRecordSizeEstimator(schema));
+    assertTrue(payloadSize > 0);
+
+    // Test sizeEstimator with hoodie metadata fields
+    schema = HoodieAvroUtils.addMetadataFields(schema);
+    hoodieRecords = SchemaTestUtil.generateHoodieTestRecords(0, 1, schema);
+    payloadSize = SpillableMapUtils.computePayloadSize(hoodieRecords.remove(0), new HoodieRecordSizeEstimator(schema));
+    assertTrue(payloadSize > 0);
+
+    // Following tests payloads without an Avro Schema in the Record
+
+    // Test sizeEstimator without hoodie metadata fields and without schema object in the payload
+    schema = SchemaTestUtil.getSimpleSchema();
+    List<IndexedRecord> indexedRecords = SchemaTestUtil.generateHoodieTestRecords(0, 1);
+    hoodieRecords =
+        indexedRecords.stream().map(r -> new HoodieRecord<>(new HoodieKey(UUID.randomUUID().toString(), "0000/00/00"),
+            new AvroBinaryTestPayload(Option.of((GenericRecord) r)))).collect(Collectors.toList());
+    payloadSize = SpillableMapUtils.computePayloadSize(hoodieRecords.remove(0), new HoodieRecordSizeEstimator(schema));
+    assertTrue(payloadSize > 0);
+
+    // Test sizeEstimator with hoodie metadata fields and without schema object in the payload
+    final Schema simpleSchemaWithMetadata = HoodieAvroUtils.addMetadataFields(SchemaTestUtil.getSimpleSchema());
+    indexedRecords = SchemaTestUtil.generateHoodieTestRecords(0, 1);
+    hoodieRecords = indexedRecords.stream()
+        .map(r -> new HoodieRecord<>(new HoodieKey(UUID.randomUUID().toString(), "0000/00/00"),
+            new AvroBinaryTestPayload(
+                Option.of(HoodieAvroUtils.rewriteRecord((GenericRecord) r, simpleSchemaWithMetadata)))))
+        .collect(Collectors.toList());
+    payloadSize = SpillableMapUtils.computePayloadSize(hoodieRecords.remove(0), new HoodieRecordSizeEstimator(schema));
+    assertTrue(payloadSize > 0);
+  }
+
+  @Test
   public void testPutAll() throws IOException, URISyntaxException {
-    SpillableRocksDBBasedMap<String, HoodieRecord> rocksDBBasedMap = new SpillableRocksDBBasedMap<>(basePath);
+    BitCaskDiskMap<String, HoodieRecord> records = new BitCaskDiskMap<>(basePath);
     List<IndexedRecord> iRecords = SchemaTestUtil.generateHoodieTestRecords(0, 100);
     Map<String, HoodieRecord> recordMap = new HashMap<>();
     iRecords.forEach(r -> {
@@ -179,30 +199,31 @@ public class TestSpillableRocksDBBasedMap extends HoodieCommonTestHarness {
       recordMap.put(key, value);
     });
 
-    rocksDBBasedMap.putAll(recordMap);
+    records.putAll(recordMap);
     // make sure records have spilled to disk
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > 0);
+    assertTrue(records.sizeOfFileOnDiskInBytes() > 0);
 
     // make sure all added records are present
-    for (Map.Entry<String, HoodieRecord> entry : rocksDBBasedMap.entrySet()) {
+    for (Map.Entry<String, HoodieRecord> entry : records.entrySet()) {
       assertTrue(recordMap.containsKey(entry.getKey()));
     }
   }
 
+  /**
+   * @na: Leaving this test here for a quick performance test
+   */
+  @Disabled
   @Test
-  public void testSimpleRemove() throws IOException, URISyntaxException {
-    SpillableRocksDBBasedMap rocksDBBasedMap = new SpillableRocksDBBasedMap(basePath);
-    List<IndexedRecord> iRecords = SchemaTestUtil.generateHoodieTestRecords(0, 100);
-    ((GenericRecord) iRecords.get(0)).get(HoodieRecord.COMMIT_TIME_METADATA_FIELD).toString();
-    List<String> recordKeys = SpillableMapTestUtils.upsertRecords(iRecords, rocksDBBasedMap);
-
-    // make sure records have spilled to disk
-    assertTrue(rocksDBBasedMap.sizeOfFileOnDiskInBytes() > 0);
-
-    List<String> deleteKeys = recordKeys.subList(0, 10);
-    for (String deleteKey : deleteKeys) {
-      assert rocksDBBasedMap.remove(deleteKey) != null;
-      assert rocksDBBasedMap.get(deleteKey) == null;
-    }
+  public void testSizeEstimatorPerformance() throws IOException, URISyntaxException {
+    // Test sizeEstimatorPerformance with simpleSchema
+    Schema schema = SchemaTestUtil.getSimpleSchema();
+    List<HoodieRecord> hoodieRecords = SchemaTestUtil.generateHoodieTestRecords(0, 1, schema);
+    HoodieRecordSizeEstimator sizeEstimator = new HoodieRecordSizeEstimator<>(schema);
+    HoodieRecord record = hoodieRecords.remove(0);
+    long startTime = System.currentTimeMillis();
+    SpillableMapUtils.computePayloadSize(record, sizeEstimator);
+    long timeTaken = System.currentTimeMillis() - startTime;
+    System.out.println("Time taken :" + timeTaken);
+    assertTrue(timeTaken < 100);
   }
 }
